+++ apps/api/src/modules/collections/dto/create-child-collection.dto.ts
@@ -1,6 +1,0 @@
-import { OmitType } from "@nestjs/mapped-types";
-import { CreateCollectionDto } from "./create-collection.dto";
-
-export class CreateChildCollectionDto extends OmitType(CreateCollectionDto, [
-  "parentCollectionId",
-] as const) {}